requirements.txt
boto3==1.36.13
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
pandas==2.2.3
pendulum==3.0.0
pyjwt==2.10.1
python-dotenv==1.0.1
requests==2.32.3